+++ analyzer/src/main/java/nl/basjes/parse/useragent/analyze/MatcherExtractAction.java
@@ -24,6 +24,7 @@ import org.antlr.v4.runtime.ParserRuleContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+@SuppressWarnings("this-escape")
 public class MatcherExtractAction extends MatcherAction {
     private static final Logger LOG = LogManager.getLogger(MatcherExtractAction.class);